code cleanup for client/service connection classes
